+++ CHANGELOG.md
@@ -376,3 +376,11 @@
     - ShopItemGridView와 Presenter 구현
       - Cart와 마찬가지로 Create를 통해 View 반환
       - Remove와 Clear, Dispose 구현
+
+## 2026-09-22
+
+    - 상점 NPC 대화(대사) 연결
+      - SO_DialogueData에 조건식과 변수 바인딩을 연결하여 대사 출력 엔진?처럼 구현
+        - ex: {itemCount == 0 ? 어서오세요! 물건 많으니 둘러보세요 : {itemCount == 1 ? '{item1}' {totalAmount}개면 합해서 <b>{totalCost}</b>(이네/네)요! : '{item1}', '{item2}' 등 다 합해서 {totalAmount}개니까 <b>{totalCost}</b>(이네/네)요!}}
+        - 기존 NPC 대화와 구분을 위해 StartShopDialogueEvent로 구분
+      - 이름과 대사를 위한 박스의 크기를 강제 갱신